requirements.txt
numpy
pandas
scikit-learn>=1.0.0
matplotlib
pyyaml
joblib

